main hat die willkürliche Grenze von 48 Farben aufgehoben (PR #13): jetzt
sind so viele Farben möglich, wie der Garnkatalog hat, also 375. Dazu
kamen ein breiteres Raster (Uint16 statt Uint8) und die Abstandsliste
anstelle der Abstandstabelle.

Zusammengeführt:

  - `farbenNeu` im Worker ruft jetzt `abstandslisteBauen` und das
    einarmige `ohneGlaettungZuordnen` – die Umbenennung aus main.
  - Der Farbregler deckt den ganzen Bereich ab. Auf einer geraden Skala
    von 2 bis 375 läge alles unter 40 Farben in den ersten Millimetern,
    deshalb wächst die Farbzahl geometrisch: Mitte bei 27, ein Viertel
    bei 7, drei Viertel bei 101. Hin und zurück ist die Umrechnung
    stabil, der Regler zittert beim Ziehen also nicht.
  - Die Knöpfe „Weniger"/„Mehr" gehen in denselben Stufen wie die
    Zahlenwahl in Schritt 2. `farbenSchritt` steht dafür nicht mehr in
    der Einstellungsseite, sondern in typen.ts – zwei verschiedene
    Schrittweiten für dieselbe Zahl wären nicht zu erklären.
  - Das Zahlenfeld ist breit genug für dreistellige Werte.

Der Kommentar bei MAX_FARBEN aus main stimmt jetzt genauer als vorher:
mehr Farben als Garne zu verlangen ergibt wirklich keinen Sinn, weil
jedes Cluster bis zur Erschöpfung des Katalogs ein eigenes Garn bekommt.

+++ src/app/schritt/muster/MusterAnsehen.tsx
@@ -415,7 +415,7 @@ export function MusterAnsehen() {
   // dem Pinselstrich, der gerade noch gezogen wird.
   const anzeigeRaster = useMemo(() => {
     if (!raster || !malSpur || malSpur.size === 0) return raster;
-    const kopie = Uint8Array.from(raster);
+    const kopie = Uint16Array.from(raster);
     for (const [feld, wert] of malSpur) kopie[feld] = wert;
     return kopie;
   }, [raster, malSpur]);
