+++ src/lib/speicher/staende.ts
@@ -45,7 +45,7 @@ export type Stand = {
 export type StandInhalt = {
   breite: number;
   hoehe: number;
-  basis: Uint8Array;
+  basis: Uint16Array;
   bearbeitung: Int16Array;
   palette: PalettenEintrag[];
 };
@@ -69,7 +69,7 @@ type Abgelegt = {
 async function vorschauBauen(
   breite: number,
   hoehe: number,
-  raster: Uint8Array,
+  raster: Uint16Array,
   palette: PalettenEintrag[],
 ): Promise<Blob | null> {
   if (typeof document === "undefined") return null;
@@ -112,9 +112,9 @@ export async function standSichern(argumente: {
   gemerkt: boolean;
   breite: number;
   hoehe: number;
-  basis: Uint8Array;
+  basis: Uint16Array;
   bearbeitung: Int16Array;
-  raster: Uint8Array;
+  raster: Uint16Array;
   palette: PalettenEintrag[];
   einstellungen: Einstellungen;
   quellbild: Blob | null;
